third_party/WebKit/LayoutTests/inspector/sources/debugger-pause/debugger-pause-on-promise-rejection-expected.txt

CONSOLE ERROR: line 35: Uncaught (in promise) Error: uncaught
CONSOLE ERROR: line 35: Uncaught (in promise) Error: uncaught
Tests that pause on promise rejection works.

=== Pausing only on uncaught exceptions ===
Set timer for test function.
Call stack:
    0) testFunction (debugger-pause-on-promise-rejection.html:35)
Paused on promise rejectionError: uncaught

=== Pausing on all exceptions ===
Set timer for test function.
Call stack:
    0) testFunction (debugger-pause-on-promise-rejection.html:34)
Paused on promise rejectionError: caught
Call stack:
    0) testFunction (debugger-pause-on-promise-rejection.html:35)
Paused on promise rejectionError: uncaught
